refactor: 에러가 나도 프로그램이 정상종료하도록 수정
